requirements.txt
openai
anthropic
mysql-connector-python
python-dotenv
tabulate
rich
fastapi
uvicorn
pytest
numpy
scipy
requests
pandas